Modification for UI on list update, board update & card delete

+++ app/Livewire/Label/LabelDelete.php
@@ -14,6 +14,9 @@ class LabelDelete extends Component
     }
 
     public function deleteLabel() {
+        if(! $this->label) {
+            abort(404, 'Label not found');
+        }
         $this->label->delete();
         $this->dispatch('label-deleted');
     }
